Move grammars back to LoopIR.py
